allow output and auth flags after commands
