+++ auth/cliopts.go
@@ -7,6 +7,9 @@ func cliopts(c *cli.Context, have map[string]string, need map[string]string) {
 		if c.GlobalIsSet(opt) {
 			have[opt] = c.GlobalString(opt)
 			delete(need, opt)
+		} else if c.IsSet(opt) {
+			have[opt] = c.String(opt)
+			delete(need, opt)
 		}
 	}
 }
